fix(#2438): update stale nexus.security refs inside moved lib/security/ files
The docstrings and runtime/TYPE_CHECKING imports within the moved
security modules still referenced the old nexus.security path,
causing ModuleNotFoundError at runtime.
